+++ example/NotifService.js
@@ -157,7 +157,7 @@ export default class NotifService {
   }
 
   cancelNotif() {
-    PushNotification.cancelLocalNotifications({id: '' + this.lastId});
+    PushNotification.cancelLocalNotification(this.lastId);
   }
 
   cancelAll() {